+++ src/shared/utils/config.js
@@ -11,6 +11,5 @@ import { isClientSide } from './isomorphy';
 import { requireWeak } from './webpack';
 
 /* eslint-disable global-require */
-// console.log('IS CLIENT SIDE', isClientSide());
-module.exports = isClientSide() ? window.CONFIG : requireWeak('config');
+export default isClientSide() ? window.CONFIG : requireWeak('config');
 /* eslint-enable global-require */