+++ src/shared/utils/redux.js
@@ -11,9 +11,7 @@ import { connect } from 'react-redux';
 import { isDevBuild } from './isomorphy';
 
 /* Auxiliary aliases. */
-module.exports.connect = connect;
-module.exports.createActions = createActions;
-module.exports.handleActions = handleActions;
+export { connect, createActions, handleActions };
 
 /**
  * Reduce multiple reducers into a single reducer from left to right.